requirements.txt
aiohttp
aiofiles
beautifulsoup4
pillow
ebooklib
piexif
motor
pyrogram
